use common extension manifest string replacer for web and desktop
